# RFC: improve initramfs image performance and efficiency via cpio reflinks