--- a/dracut-bash-completion.sh
+++ b/dracut-bash-completion.sh
@@ -32,7 +32,8 @@ _dracut() {
             --local --hostonly --no-hostonly --fstab --help --bzip2 --lzma
             --xz --zstd --no-compress --gzip --list-modules --show-modules --keep
             --printsize --regenerate-all --noimageifnotneeded --early-microcode
-            --no-early-microcode --print-cmdline --reproducible --uefi'
+            --no-early-microcode --print-cmdline --reproducible --uefi
+            --cpio-reflink'
         [ARG]='-a -m -o -d -I -k -c -L --kver --add --force-add --add-drivers
             --omit-drivers --modules --omit --drivers --filesystems --install
             --fwdir --libdirs --fscks --add-fstab --mount --device --nofscks