--- a/dracut.8.asc
+++ b/dracut.8.asc
@@ -531,6 +531,15 @@ will not be able to boot.
     Specifies the kernel image, which to include in the UEFI executable. The default is
     _/lib/modules/<KERNEL-VERSION>/vmlinuz_ or _/boot/vmlinuz-<KERNEL-VERSION>_
 
+**--cpio-reflink**::
+    Attempt to use the GNU cpio --reflink option, which optimizes archive creation for
+    copy-on-write filesystems by using the copy_file_range(2) syscall. When specified,
+    initramfs archives are also padded to ensure optimal data alignment for extent
+    sharing. To retain reflink data deduplication benefits, this should be used
+    alongside the **--no-compress** and **--no-strip** parameters, with initramfs
+    source files, **--tmpdir** staging area and destination all on the same
+    copy-on-write capable filesystem.
+
 ENVIRONMENT
 -----------