--- a/test/Makefile.testdir
+++ b/test/Makefile.testdir
@@ -2,11 +2,11 @@
 
 all:
 	@$(MAKE) -s --no-print-directory -C ../.. all
-	@V=$(V) basedir=../.. testdir=../ ./test.sh --all
+	@V=$(V) basedir=$(realpath ../..) testdir="$(realpath ../)" ./test.sh --all
 setup:
 	@$(MAKE) --no-print-directory -C ../.. all
-	@basedir=../.. testdir=../ ./test.sh --setup
+	@basedir="$(realpath ../..)" testdir="$(realpath ../)" ./test.sh --setup
 clean:
-	@basedir=../.. testdir=../ ./test.sh --clean
+	@basedir="$(realpath ../..)" testdir="$(realpath ../)" ./test.sh --clean
 run:
-	@basedir=../.. testdir=../ ./test.sh --run
+	@basedir="$(realpath ../..)" testdir="$(realpath ../)" ./test.sh --run