--- a/test/TEST-62-CPIO/Makefile
+++ b/test/TEST-62-CPIO/Makefile
@@ -1,0 +1,1 @@
+-include ../Makefile.testdir