… test/TEST-62-CPIO/test.sh