This patchset attempts to speed up initramfs generation for some common (Btrfs / XFS) setups by having Dracut make heavier use of reflinks (AKA copy-on-write clones) during initramfs generation. A good portion of an uncompressed+unstripped initramfs image is duplicate data, which really shouldn't need to be shuffled around when on the same COW clone capable FS.

Dracut already uses `cp --reflink=auto` when shuffling most things into the temporary staging area. This patchset adds a new `--cpio-reflink` flag which sees Dracut call GNU cpio using the new `--reflink` and `--chain` parameters. Support for these new parameters can be found in my cpio repository at https://github.com/ddiss/cpio/tree/copy_file_range_2_13 .
As XFS and Btrfs require filesystem alignment for extent sharing, a new `padcpio` utility is added to provide padding of file data to a block-size boundary.

This allows for:
* improved space efficiency
  * initramfs creation avoids data duplication
* improved performance
  * initramfs image needn't be stripped / compressed / decompressed

The following caveats would be present for dracut to successfully use reflink (otherwise fallback to read/write):
* root, boot and dracut staging (/var/tmp) exist on the same Btrfs or XFS filesystem
* paths don't have nocow flags set
* boot performance may be negatively affected by fragmentation, but that should be compensated by the removal of compression / decompression

Benchmarks are still ongoing, but my initial results show:
```
---------------------------------+----------+----------+-----------
     Benchmark                   |  Before  |  After   |  Change
---------------------------------+----------+----------+-----------
Dracut create image runtime      |  8.452s  |  7.635s  |  -9.666%
---------------------------------+----------+----------+-----------
initramfs data (fiemap)          |          |          |
- total                          | 12894208 | 34009088 |  +163.7%
- shared (dedup)                 |     0    | 24068096 |
- exclusive                      | 12894208 |  9940992 |  -22.90%
---------------------------------+----------+----------+-----------
QEMU cold boot to Dracut init    |  3.208s  |  2.850s  |  -11.15%
---------------------------------+----------+----------+-----------
```
See https://github.com/dracutdevs/dracut/issues/1141#issuecomment-792111392 for details.

RFC: The top commit of this patchset shouldn't be merged until GNU `cpio --reflink` support is upstream. I believe all other changes are read.

Fixes: https://github.com/dracutdevs/dracut/issues/1141