@ddiss 2021-03-11
> So, the whole thing is most interesting for initramfs images stored on xfs/btrfs for qemu booting, as moving it to EFI vfat will remove the dedup feature. Boot times for bare metal would also increase due to non-compression.
> 
> Did I understand this correctly?

Yes, correct, it's most beneficial for environments where initramfs source files, Dracut staging area and destination image are all on the same XFS or Btrfs filesystem. If the Dracut destination is EFI vfat then duplication is expected, although it would still allow for `$initdir/*`->`${DRACUT_TMPDIR}/initramfs.img` cpio data copy I/O to be reduced.