@ddiss 2021-03-11
> I am sorry, you have to rebase due to upstream changes in `Makefile`.

No problem :smiley: . I've made some extra changes to use statfs reported optimal alignment for padding, rather than the hardcoded `4K`. I'll force push here along with @johannbg 's suggested changes when done testing.