@ddiss 2021-03-12
> @ddiss can you give us a link to your cpio --reflink upstream PR, looking upstream and based on lack of activity there,it looks like it can literally take years before your PR gets merged and is included in a new release.

They've now been submitted via https://savannah.gnu.org/patch/index.php?10044 and the upstream mailing list:
https://lists.gnu.org/archive/html/bug-cpio/2021-02/index.html