@ddiss 2021-03-15
Changes since previous version:
- rebase against current master branch
- use statfs to determine optimal padding alignment
- address @johannbg 's review feedback
- add minor test cleanup: d139603 
- improve man-page documentation