@ddiss 2021-03-15
Changes since previous version:
- rebase against current master branch
- apply `astyle` indent rules to padcpio