@ddiss 2021-03-15
Changes since previous version:
- add `padcpio` to dracut.spec