@ddiss 2021-03-15
Changes since previous version:
- change TEST-62-CPIO commit message to make Commisery happy