@haraldh 2021-03-30
I will merge this, if cpio upstream supports it.