@haraldh 2021-05-19
> After seeing Rust's native support for `copy_file_range()` via `fs::copy()`, I went ahead and wrote a prototype `dracut-cpio` utility:
> https://github.com/ddiss/dracut/tree/dracut_cpio

Here is my `cp` version in rust for `dracut-install` with `hole` support:
https://github.com/haraldh/dracut-install/blob/master/src/file/mod.rs#L228