@haraldh 2021-05-19
On a funny side note, part of `fs::copy()` was [one of my first rust project contributions](https://github.com/rust-lang/rust/pull/58803/files) :)