@ddiss 2021-05-19
> Here is my (outdated) cp version in rust for dracut-install with sparse support:
> https://github.com/haraldh/dracut-install/blob/master/src/file/mod.rs#L228

Nice! I wasn't aware of your work in this area. I'd also be happy to integrate the dracut-cpio functionality into dracut-install.rs directly if you plan on continuing in that direction.