@ddiss 2021-05-27
> After seeing Rust's native support for `copy_file_range()` via `fs::copy()`, I went ahead and wrote a prototype `dracut-cpio` utility:
> https://github.com/ddiss/dracut/tree/dracut_cpio

I've updated the dracut-cpio patch to the point where it can now be used for proper testing as a swap-in replacement for GNU cpio. Like my GNU cpio patchset, it supports cpio reflinks and with data segment alignment. Data segment alignment is handled in a much cleaner fashion by injecting extra zeros into the filename field (rather than adding extra pad files to initramfs).
GNU cpio and kernel seem to handle filename zero-injection just fine, but I need to do some further testing and benchmarking.